third_party/blink/web_tests/external/wpt/editing/other/insertparagraph-in-child-of-head.tentative_designMode=on_white-space=normal-expected.txt

This is a testharness.js-based test.
[FAIL] insertParagraph in empty <title style="display:block;white-space:normal"> should not split it
  assert_in_array: 0-2 collapsible linefeed(s) should be inserted value "<head><title></title></head><body><br><div><br></div></body>" not in array ["<head><title></title></head><body><br></body>", "<head><title>\\n</title></head><body><br></body>", "<head><title>\\n\\n</title></head><body><br></body>"]
[FAIL] insertParagraph in <title style="display:block;white-space:normal"> containing text should not split it
  assert_in_array: 0-1 collapsible linefeed should be inserted value "<head><title>ab</title><title style=\\"display:block;white-space:normal\\">cd</title></head><body><br></body>" not in array ["<head><title>abcd</title></head><body><br></body>", "<head><title>ab\\ncd</title></head><body><br></body>"]
[FAIL] insertParagraph in empty <style style="display:block;white-space:normal"> should not split it
  assert_in_array: 0-2 collapsible linefeeds should be inserted value "<head><title>iframe</title><style></style></head><body><br><div><br></div></body>" not in array ["<head><title>iframe</title><style></style></head><body><br></body>", "<head><title>iframe</title><style>\\n</style></head><body><br></body>", "<head><title>iframe</title><style>\\n\\n</style></head><body><br></body>"]
[FAIL] insertParagraph in <style style="display:block;white-space:normal"> containing text should not split it
  assert_in_array: 0-1 collapsible linefeed should be inserted value "<head><title>iframe</title><style>ab</style><style style=\\"display:block;white-space:normal\\">cd</style></head><body><br></body>" not in array ["<head><title>iframe</title><style>abcd</style></head><body><br></body>", "<head><title>iframe</title><style>ab\\ncd</style></head><body><br></body>"]
[FAIL] insertParagraph in empty <script style="display:block;white-space:normal"> should not split it
  assert_in_array: 0-2 collapsible linefeeds should be inserted value "<head><title>iframe</title><script></script></head><body><br><div><br></div></body>" not in array ["<head><title>iframe</title><script></script></head><body><br></body>", "<head><title>iframe</title><script>\\n</script></head><body><br></body>", "<head><title>iframe</title><script>\\n\\n</script></head><body><br></body>"]
[FAIL] insertParagraph in <script style="display:block;white-space:normal"> containing text should not split it
  assert_in_array: 0-1 linefeed should be inserted value "<head><title>iframe</title><script>// ab</script><script style=\\"display:block;white-space:normal\\">// cd</script></head><body><br></body>" not in array ["<head><title>iframe</title><script>// ab// cd</script></head><body><br></body>", "<head><title>iframe</title><script>// ab\\n// cd</script></head><body><br></body>"]
[FAIL] insertParagraph in empty <div style="white-space:normal"> in the <head> should split the <div>
  assert_equals: The <div> should be split expected "<head><title>iframe</title><div style=\\"white-space:normal\\"><br></div><div style=\\"white-space:normal\\"><br></div></head><body><br></body>" but got "<head><title>iframe</title><div style=\\"white-space:normal\\"></div></head><body><br><div><br></div></body>"
Harness: the test ran to completion.
